@@ -9,25 +9,28 @@
 		"build-vendors": "webpack --mode production --config ./vendor-webpack.config.js"
 	},
 	"dependencies": {
-		"@angular/common": "~6.0.0",
-		"@angular/compiler": "~6.0.0",
-		"@angular/core": "~6.0.0",
-		"@angular/forms": "~6.0.0",
-		"@angular/platform-browser": "~6.0.0",
-		"@angular/platform-browser-dynamic": "~6.0.0",
-		"@angular/router": "~6.0.0",
-		"@angular/material": "~6.0.0",
-		"@angular/cdk": "~6.0.0",
-		"@angular/animations": "~6.0.0",
+		"@angular/common": "~7.1.0",
+		"@angular/compiler": "~7.1.0",
+		"@angular/core": "~7.1.0",
+		"@angular/forms": "~7.1.0",
+		"@angular/platform-browser": "~7.1.0",
+		"@angular/platform-browser-dynamic": "~7.1.0",
+		"@angular/router": "~7.1.0",
+		"@angular/material": "~7.1.0",
+		"@angular/cdk": "~7.1.0",
+		"@angular/animations": "~7.1.0",
 		"core-js": "2.5.7",
-		"rxjs": "6.2.2",
+		"rxjs": "6.3.3",
 		"zone.js": "0.8.26",
-		"chart.js": "*"
+		"chart.js": "*",
+		"ngx-scrollbar": "*",
+		"leaflet": "*"
 	},
 	"devDependencies": {
 		"@types/node": "10.5.2",
 		"@types/chart.js": "*",
-		"typescript": "~3.0.0",
+		"@types/leaflet": "",
+		"typescript": "~3.1.6",
 		"webpack": "~4.0.0",
 		"webpack-cli": "3.1.0",
 		"webpack-dev-server": "3.1.4",
